Don't use grep -E (switch to _egrep_o); avoid using head -c.
Dropping the "head -c -2" call does mean the method returns trailing
junk at the end of the JSON fragment it returns however that will
already be the case if the JSON from the server does not have the
"result" element last. In my testing the metaname servers do not appear
to be consistent with the order of elements so this does happen.

Thankfully it doesn't actually matter since we just use dumb string
operations to pull out chunks of information from it.